refactory main tables
